Remove RawContent property
The RawContent property is in the Powershell Invoke-WebRequest comdlet,
but .Net's HttpResponse object does not return the raw resonse at all. I
could construct it from the fields returned in HttpResponse, but it
would be a fabrication, and I'd rather be honest about what I got, than
attempt to construct something I don't actually have access to.

This also gave me the ability to tighten up the output extraction.